[hydrostatic] Added -Wconversion flag
Removed all implicit casts. This was done because a bug was
discovered in the hydrostatic module, due to implicit casting:
a string was provided in lieu of a Point object but since the
Point class has a constructor taking a string as parameter, the
string was silently converted to a Point, its coordinates being
initialized at zero.
The -Wconversion flag is supposed to stop these kinds of problems
from occruring in the future.

+++ code/numeric/inc/pairwise_sum.hpp
@@ -19,7 +19,7 @@ namespace sum
         if (x.empty()) return T()-T();
         if (x.size() == 1) return x.front();
         size_t n = x.size();
-        const size_t p = floor(n/2.+0.5);
+        const size_t p = (size_t)floor(n/2.+0.5);
         size_t last = 0;
         for (size_t i = 0 ; i < p ; ++i)
         {
